push

+++ ArrayBilangan22.java
@@ -1,6 +1,6 @@
 public class ArrayBilangan22 {
     public static void main(String[] args) {
-        int[] bil = {5.0, 12867, 7.5, 2000000};
+        double[] bil = {5.0, 12867, 7.5, 2000000};
 
         for (int i = 0; i <= 4; i++) {
             System.out.println(bil[i]);
